distribuidoras_energia_analise/pyproject.toml
[project]
name = "distribuidoras-energia-analise"
version = "0.1.0"
description = ""
authors = [
    {name = "rafael",email = "devrafaelalvarenga@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg (>=0.31.0,<0.32.0)",
    "pandas (>=3.0.0,<4.0.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.0,<0.16.0)",
    "black (>=26.1.0,<27.0.0)"
]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
